+++ src/main/kotlin/io/kuzzle/sdk/coreClasses/exceptions/KuzzleException.kt
@@ -1,4 +1,4 @@
-package io.kuzzle.sdk.protocolTest.coreClasses.exceptions
+package io.kuzzle.sdk.coreClasses.exceptions
 
 /**
  * Root of all Kuzzle exceptions.
