+++ src/main/kotlin/io/kuzzle/sdk/coreClasses/exceptions/KuzzleExceptionCode.kt
@@ -1,4 +1,4 @@
-package io.kuzzle.sdk.protocolTest.coreClasses.exceptions
+package io.kuzzle.sdk.coreClasses.exceptions
 
 enum class KuzzleExceptionCode {
   MISSING_REQUESTID(0, "Missing field requestId"), MISSING_QUERY(400, "You must provide a query"), NOT_CONNECTED(500, "Not connected."), CONNECTION_LOST(500, "Connection lost"), WRONG_VOLATILE_TYPE(
