+++ src/main/kotlin/io/kuzzle/sdk/protocol/AbstractProtocol.kt
@@ -1,6 +1,6 @@
-package io.kuzzle.sdk.protocolTest.protocol
+package io.kuzzle.sdk.protocol
 
-import io.kuzzle.sdk.protocolTest.events.EventManager
+import io.kuzzle.sdk.events.EventManager
 import java.util.concurrent.ConcurrentHashMap
 
 abstract class AbstractProtocol: EventManager() {
